// hr-frontend/src/lib/api/generated/model/updateOrganizationUnitRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface UpdateOrganizationUnitRequest {
  /**
     * @minLength 0
     * @maxLength 50
     */
  code: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  nameEn: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  nameAr: string;
  unitTypeId: number;
  parentOrgUnitId?: number;
  description?: string;
  descriptionAr?: string;
  startDate: string;
  endDate?: string;
  active?: boolean;
}